+++ src/components/DataDisplay/AvatarAppBar/AvatarAppBar.tsx
@@ -1,0 +1,18 @@
+import { Avatar, AvatarProps } from "@mui/material";
+import { dark } from "@/constants/colors";
+
+const AvatarAppBar = ({ sx, ...props }: AvatarProps) => (
+  <Avatar
+    variant="rounded"
+    sx={{
+      "& svg": {
+        color: dark?.palette?.text?.disabled,
+      },
+      backgroundColor: dark?.palette?.grey?.[100],
+      color: dark?.palette?.text?.primary,
+    }}
+    {...props}
+  />
+);
+
+export default AvatarAppBar;