+++ src/components/DataDisplay/AvatarAppBar/index.ts
@@ -1,0 +1,4 @@
+import AvatarAppBar from "./AvatarAppBar";
+
+export default AvatarAppBar;
+export * from "./AvatarAppBar";