+++ src/components/Layout/Backoffice/stories/Backoffice.stories.tsx
@@ -33,7 +33,36 @@ const secondaryMenuItems = [
 const Template: StoryFn<typeof Backoffice> = (args) => (
   <Box sx={{ height: "100%", width: "100%" }}>
     <Backoffice
-      AppBar={<AppBar />}
+      AppBar={
+        <AppBar
+          actionProps={{
+            children: "Action",
+          }}
+        />
+      }
+      Sidebar={
+        <NavigationMenu
+          items={menuItems}
+          secondaryMenu={{
+            items: secondaryMenuItems,
+            label: "Settings",
+            startIcon: "⚙️",
+          }}
+        />
+      }
+      Main={
+        <Box p={3}>
+          <Typography variant="h1">This is main</Typography>
+        </Box>
+      }
+      {...args}
+    />
+  </Box>
+);
+
+const WithoutAppBarTemplate: StoryFn<typeof Backoffice> = (args) => (
+  <Box sx={{ height: "100%", width: "100%" }}>
+    <Backoffice
       Sidebar={
         <NavigationMenu
           items={menuItems}
@@ -56,6 +85,9 @@ const Template: StoryFn<typeof Backoffice> = (args) => (
 
 export const Basic = Template.bind({});
 Basic.args = {};
+
+export const WithoutAppBar = WithoutAppBarTemplate.bind({});
+WithoutAppBar.args = {};
 
 export default {
   component: Backoffice,