+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.0.0
+## v2.0.0-beta.1
 - **[feat]** - update theme colors
 - **[feat]** - add new`Backoffice` component
 - **[feat]** - add new`AppBar` component
 - **[feat]** - Add new `Logo`
 - **[feat]** - Add new `Logo` for supplier
 - **[feat]** - Add new `TextFieldAppBar` component
-- **[reactor]** - adapte `NavigationMenu` component for new `AppBar` component
+- **[feat]** - Add new `ActionAppBar` component
+- **[feat]** - Add new `AvatarAppBar` component
+- **[feat]** - adapte `NavigationMenu` component for new `AppBar` component
 - **[chore]** - Improve module augmentation declaration
+- **[docs]** - Add new story for `AppBar` component
+- **[docs]** - Add new story for `Backoffice` component